… plugins/textual/importtext_procedure.h
… sql/sql_expression.cc
… sql/sql_expression.h
… sql/sql_expression_operations.cc
… testing/decomposition_unit_test.js
… testing/testing.mk